requirements.txt
jsonpickle==1.3
#aws-xray-sdk==2.4.3
boto3==1.14.20
psycopg2-binary==2.8.6
cryptography>=2.1
credstash==1.10.0